Use ArmDev 2.0
- test Python 2.6
- test Django 1.5 and 1.6
- drop Lettuce req and installed_apps (never used anyway)
