+++ .travis.yml
@@ -1,12 +1,17 @@
 language: python
 python:
- - 2.7
+  - 2.6
+  - 2.7
 install:
-  - ./.travis_setup
+  - pip install -r requirements/dev.txt
+  - pip install .
+  - pip install Django==$DJANGO_VERSION
 env:
-  - SKIP_COVERAGE=1 SKIP_INSTALL=1 DJANGO_VERSION=1.4.5
-  - SKIP_COVERAGE=1 SKIP_INSTALL=1 DJANGO_VERSION=1.3.3
-script: fab test
+  - DJANGO_VERSION=1.6.11
+  - DJANGO_VERSION=1.5.12
+  - DJANGO_VERSION=1.4.20
+  - DJANGO_VERSION=1.3.7
+script: python -m armstrong.dev.dev_django test
 notifications:
   email: false
   irc:
