--- .travis_setup
@@ -1,4 +1,0 @@
-#!/bin/bash
-pip install -i http://armstrong.github.com/pypi.armstrongcms.org/index/ -r requirements/dev.txt
-pip install .
-pip install Django==$DJANGO_VERSION
