+++ env_settings.py
@@ -1,0 +1,20 @@
+from armstrong.dev.default_settings import *
+
+
+INSTALLED_APPS.extend([
+    'django.contrib.admin',
+    'django.contrib.auth',
+    'django.contrib.contenttypes',
+    'django.contrib.sessions',
+    'armstrong.hatband.tests.hatband_support'
+])
+MIDDLEWARE_CLASSES = (
+    'django.middleware.common.CommonMiddleware',
+    'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
+    'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'django.contrib.messages.middleware.MessageMiddleware'
+)
+
+STATIC_URL = "/static/"
+ROOT_URLCONF = 'armstrong.hatband.tests.hatband_support.urls'
